- Update to node 16 and Debian bullseye (with fallback to node 14 for older incompatible SDKs)
- Handle some older SDK versions which output lockdown sniffing to stdout instead of stderr
- Rewrote the config argv parsing logic, to make it behave slightly more sanely
- Fixed some deadlock issues, e.g. adding some timeouts on task ready (see #40) or slog streams not closing properly
- Capture client storage and slog file (should help track some transient seg faults in the solo)
- Automatically capture the state of the client and chain if an error occurs (see #39)
- Background the compression of the state directories after snapshotting them (overlayfs supports CoW). Closes #39 
- Avoid resetting the whole `agoric-servers` project in `local-chain` tests. Removes loadgen project `git` dependency.
- Elide long lines from the chain or solo output (improves github actions perf, see https://github.com/Agoric/agoric-sdk/pull/4113)

+++ runner/lib/tasks/local-chain.js
@@ -10,6 +10,7 @@ import {
   childProcessReady,
 } from '../helpers/child-process.js';
 import BufferLineTransform from '../helpers/buffer-line-transform.js';
+import ElidedBufferLineTransform from '../helpers/elided-buffer-line-transform.js';
 import { PromiseAllOrErrors, tryTimeout } from '../helpers/async.js';
 import { fsStreamReady } from '../helpers/fs.js';
 import { asBuffer, whenStreamSteps } from '../helpers/stream.js';
@@ -18,6 +19,7 @@ import {
   getChildMatchingArgv,
   wrapArgvMatcherIgnoreEnvShebang,
   getConsoleAndStdio,
+  cleanAsyncIterable,
 } from './helpers.js';
 import { makeGetEnvInfo } from './shared-env-info.js';
 import { makeLoadgenTask } from './shared-loadgen.js';
@@ -27,7 +29,10 @@ const pipeline = promisify(pipelineCallback);
 const stateDir = '_agstate/agoric-servers';
 const keysDir = '_agstate/keys';
 const profileName = 'local-chain';
+const CLIENT_PORT = 8000;
 const CHAIN_PORT = 26657;
+const clientStateDir = `${stateDir}/${profileName}-${CLIENT_PORT}`;
+const chainStateDir = `${stateDir}/${profileName}-${CHAIN_PORT}`;
 const CHAIN_ID = 'agoric';
 const GAS_ADJUSTMENT = '1.2';
 const CENTRAL_DENOM = 'urun';
@@ -54,7 +59,6 @@ const chainArgvMatcher = (argv) =>
  * @param {import("child_process").spawn} powers.spawn Node.js spawn
  * @param {import("fs/promises")} powers.fs Node.js promisified fs object
  * @param {import("../helpers/fs.js").MakeFIFO} powers.makeFIFO Make a FIFO file readable stream
- * @param {import("../helpers/fs.js").FindByPrefix} powers.findDirByPrefix
  * @param {import("../helpers/procsfs.js").GetProcessInfo} powers.getProcessInfo
  * @param {import("./types.js").SDKBinaries} powers.sdkBinaries
  * @returns {import("./types.js").OrchestratorTasks}
@@ -63,7 +67,6 @@ const chainArgvMatcher = (argv) =>
 export const makeTasks = ({
   spawn: cpSpawn,
   fs,
-  findDirByPrefix,
   makeFIFO,
   getProcessInfo,
   sdkBinaries,
@@ -86,8 +89,12 @@ export const makeTasks = ({
     }
   };
 
-  /** @param {import("./types.js").TaskBaseOptions & {config?: {reset?: boolean}}} options */
-  const setupTasks = async ({ stdout, stderr, config: { reset } = {} }) => {
+  /** @param {import("./types.js").TaskBaseOptions & {config?: {reset?: boolean, withMonitor?: boolean}}} options */
+  const setupTasks = async ({
+    stdout,
+    stderr,
+    config: { reset, withMonitor } = {},
+  }) => {
     const { console, stdio } = getConsoleAndStdio(
       'setup-tasks',
       stdout,
@@ -100,17 +107,21 @@ export const makeTasks = ({
 
     console.log('Starting');
 
+    if (withMonitor !== false) {
+      if (reset) {
+        console.log('Resetting chain node');
+        await childProcessDone(
+          printerSpawn('rm', ['-rf', chainStateDir], { stdio }),
+        );
+      }
+    }
+
     if (reset) {
-      console.log('Resetting chain node and client state');
-      await childProcessDone(printerSpawn('rm', ['-rf', stateDir], { stdio }));
+      console.log('Resetting client state');
       await childProcessDone(
-        printerSpawn('git', ['checkout', '--', stateDir], {
-          stdio,
-        }),
+        printerSpawn('rm', ['-rf', clientStateDir], { stdio }),
       );
     }
-    await childProcessDone(printerSpawn('agoric', ['install'], { stdio }));
-
     console.log('Done');
   };
 
@@ -131,6 +142,7 @@ export const makeTasks = ({
 
     const chainEnv = Object.create(process.env);
     chainEnv.SLOGFILE = slogFifo.path;
+    chainEnv.CHAIN_PORT = `${CHAIN_PORT}`;
 
     const launcherCp = printerSpawn(
       'agoric',
@@ -144,7 +156,10 @@ export const makeTasks = ({
     const chainDone = childProcessDone(launcherCp, {
       ignoreExitCode: true,
     }).then((code) => {
-      if (code !== 0 && (!stopped || code !== 98)) {
+      const wasStopped = stopped;
+      stopped = true;
+
+      if (code !== 0 && (!wasStopped || code !== 98)) {
         return Promise.reject(
           new Error(`Chain exited with non-zero code: ${code}`),
         );
@@ -152,14 +167,23 @@ export const makeTasks = ({
       return 0;
     });
 
-    chainDone.then(
-      () => console.log('Chain exited successfully'),
-      (error) => console.error('Chain exited with error', error),
-    );
+    chainDone
+      .then(
+        () => console.log('Chain exited successfully'),
+        (error) => console.error('Chain exited with error', error),
+      )
+      .finally(() => {
+        if (slogFifo.pending) {
+          slogLines.end();
+          slogFifo.close();
+        }
+      });
 
-    launcherCp.stdout.pipe(stdio[1], { end: false });
+    const launcherElidedOutput = new ElidedBufferLineTransform();
+    launcherCp.stdout.pipe(launcherElidedOutput);
+    launcherElidedOutput.pipe(stdio[1], { end: false });
     const [chainStarted, firstBlock, outputParsed] = whenStreamSteps(
-      launcherCp.stdout,
+      launcherElidedOutput,
       [
         { matcher: chainStartRE },
         { matcher: chainBlockBeginRE, resultIndex: -1 },
@@ -184,21 +208,15 @@ export const makeTasks = ({
 
         console.log('Chain running');
 
-        const [storageLocation, processInfo] = await PromiseAllOrErrors([
-          chainStarted.then(findDirByPrefix),
-          getProcessInfo(
-            /** @type {number} */ (launcherCp.pid),
-          ).then((launcherInfo) =>
-            getChildMatchingArgv(launcherInfo, chainArgvMatcher),
-          ),
-        ]);
-
+        const processInfo = await getProcessInfo(
+          /** @type {number} */ (launcherCp.pid),
+        ).then((launcherInfo) =>
+          getChildMatchingArgv(launcherInfo, chainArgvMatcher),
+        );
         const stop = () => {
-          stopped = true;
-          process.kill(processInfo.pid);
-          if (slogFifo.pending) {
-            slogLines.end();
-            slogFifo.close();
+          if (!stopped) {
+            stopped = true;
+            process.kill(processInfo.pid);
           }
         };
 
@@ -206,10 +224,8 @@ export const makeTasks = ({
           stop,
           done,
           ready,
-          slogLines: {
-            [Symbol.asyncIterator]: () => slogLines[Symbol.asyncIterator](),
-          },
-          storageLocation,
+          slogLines: cleanAsyncIterable(slogLines),
+          storageLocation: chainStateDir,
           processInfo,
         });
       },
@@ -223,7 +239,7 @@ export const makeTasks = ({
   };
 
   /** @param {import("./types.js").TaskBaseOptions} options */
-  const runClient = async ({ stdout, stderr }) => {
+  const runClient = async ({ stdout, stderr, timeout = 60 }) => {
     const { console, stdio } = getConsoleAndStdio('client', stdout, stderr);
     const printerSpawn = makePrinterSpawn({
       spawn,
@@ -232,22 +248,18 @@ export const makeTasks = ({
 
     console.log('Starting client');
 
-    const portNum = 8000;
-
-    const agServer = `${stateDir}/${profileName}-${portNum}`;
-
-    const gciFile = `${stateDir}/${profileName}-${CHAIN_PORT}/config/genesis.json.sha256`;
+    const gciFile = `${chainStateDir}/config/genesis.json.sha256`;
 
     if (!(await fsExists(gciFile))) {
       throw new Error('Chain not running');
     }
 
     // Initialize the solo directory and key.
-    if (!(await fsExists(agServer))) {
+    if (!(await fsExists(clientStateDir))) {
       await childProcessDone(
         printerSpawn(
           sdkBinaries.agSolo,
-          ['init', agServer, `--webport=${portNum}`],
+          ['init', clientStateDir, `--webport=${CLIENT_PORT}`],
           {
             stdio,
           },
@@ -258,8 +270,8 @@ export const makeTasks = ({
     const rpcAddr = `localhost:${CHAIN_PORT}`;
 
     const soloAddr = (
-      await fs.readFile(`${agServer}/ag-cosmos-helper-address`, 'utf-8')
-    ).trimRight();
+      await fs.readFile(`${clientStateDir}/ag-cosmos-helper-address`, 'utf-8')
+    ).trimEnd();
 
     const keysSharedArgs = [
       `--home=${keysDir}`,
@@ -292,7 +304,7 @@ export const makeTasks = ({
           `--gas-adjustment=${GAS_ADJUSTMENT}`,
           '--broadcast-mode=block',
           '--yes',
-          `local-solo-${portNum}`,
+          `local-solo-${CLIENT_PORT}`,
           soloAddr,
         ],
         // Then send it some coins.
@@ -351,18 +363,27 @@ export const makeTasks = ({
     }
 
     // Connect to the chain.
-    const gci = (await fs.readFile(gciFile, 'utf-8')).trimRight();
+    const gci = (await fs.readFile(gciFile, 'utf-8')).trimEnd();
     await childProcessDone(
       printerSpawn(
         sdkBinaries.agSolo,
         ['set-gci-ingress', `--chainID=${CHAIN_ID}`, gci, rpcAddr],
-        { stdio, cwd: agServer },
+        { stdio, cwd: clientStateDir },
       ),
     );
 
+    const slogFifo = await makeFIFO('client.slog');
+    const slogReady = fsStreamReady(slogFifo);
+    const slogLines = new BufferLineTransform();
+    const slogPipeResult = pipeline(slogFifo, slogLines);
+
+    const clientEnv = Object.create(process.env);
+    clientEnv.SOLO_SLOGFILE = slogFifo.path;
+
     const soloCp = printerSpawn(sdkBinaries.agSolo, ['start'], {
       stdio: ['ignore', 'pipe', stdio[2]],
-      cwd: agServer,
+      cwd: clientStateDir,
+      env: clientEnv,
       detached: true,
     });
 
@@ -374,14 +395,23 @@ export const makeTasks = ({
     const soloCpReady = childProcessReady(soloCp);
     const clientDone = childProcessDone(soloCp, { ignoreKill });
 
-    clientDone.then(
-      () => console.log('Client exited successfully'),
-      (error) => console.error('Client exited with error', error),
-    );
+    clientDone
+      .then(
+        () => console.log('Client exited successfully'),
+        (error) => console.error('Client exited with error', error),
+      )
+      .finally(() => {
+        if (slogFifo.pending) {
+          slogLines.end();
+          slogFifo.close();
+        }
+      });
 
-    soloCp.stdout.pipe(stdio[1], { end: false });
+    const soloElidedOutput = new ElidedBufferLineTransform();
+    soloCp.stdout.pipe(soloElidedOutput);
+    soloElidedOutput.pipe(stdio[1], { end: false });
     const [clientStarted, walletReady, outputParsed] = whenStreamSteps(
-      soloCp.stdout,
+      soloElidedOutput,
       [
         { matcher: clientSwingSetReadyRE, resultIndex: -1 },
         { matcher: clientWalletReadyRE, resultIndex: -1 },
@@ -391,34 +421,50 @@ export const makeTasks = ({
       },
     );
 
-    const done = PromiseAllOrErrors([outputParsed, clientDone]).then(() => {});
+    const done = PromiseAllOrErrors([
+      slogPipeResult,
+      outputParsed,
+      clientDone,
+    ]).then(() => {});
 
-    try {
-      await soloCpReady;
-      await clientStarted;
+    const ready = PromiseAllOrErrors([walletReady, slogReady]).then(() => {});
 
-      console.log('Client running');
+    return tryTimeout(
+      timeout * 1000,
+      async () => {
+        await soloCpReady;
+        await clientStarted;
 
-      const stop = () => {
-        ignoreKill.signal = 'SIGTERM';
-        soloCp.kill(ignoreKill.signal);
-      };
+        console.log('Client running');
 
-      return harden({
-        stop,
-        done,
-        ready: walletReady,
-      });
-    } catch (err) {
-      // Avoid unhandled rejections for promises that can no longer be handled
-      Promise.allSettled([done, clientStarted, walletReady]);
-      soloCp.kill();
-      throw err;
-    }
+        const processInfo = await getProcessInfo(
+          /** @type {number} */ (soloCp.pid),
+        );
+
+        const stop = () => {
+          ignoreKill.signal = 'SIGTERM';
+          soloCp.kill(ignoreKill.signal);
+        };
+
+        return harden({
+          stop,
+          done,
+          ready,
+          slogLines: cleanAsyncIterable(slogLines),
+          storageLocation: clientStateDir,
+          processInfo,
+        });
+      },
+      async () => {
+        // Avoid unhandled rejections for promises that can no longer be handled
+        Promise.allSettled([done, clientStarted, walletReady]);
+        soloCp.kill();
+      },
+    );
   };
 
   return harden({
-    getEnvInfo: makeGetEnvInfo({ spawn }),
+    getEnvInfo: makeGetEnvInfo({ spawn, sdkBinaries }),
     setupTasks,
     runChain,
     runClient,
